Move `main` to the extract crate, build `tools` as an addon
The point of providing an `main` function and using `directory` is so people can re-use the main function. Simply deriving the module trait on structs automatically adds them to the tool, no code required.

Other projects using this as a base *likely* don't want to use Froglight-specific codegen, so the codegen can be written as it's own extension.